More updates
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@20322 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/VMMR0/HWSVMR0.cpp
@@ -1776,7 +1776,7 @@ ResumeExecution:
                             &&  cbOp == 6)
                         {
                             RTGCPTR  oldEip   = pCtx->eip;
-                            RTGCPTR  GCPtrTpr = Cpu.param1.disp32;
+                            RTGCPTR  GCPtrTpr = (uint32_t)Cpu.param1.disp32;
                             uint32_t uMmioReg = Cpu.param2.base.reg_gen;
 
                             /* Found:
@@ -1791,7 +1791,7 @@ ResumeExecution:
                                 &&  Cpu.pCurInstr->opcode == OP_MOV
                                 &&  Cpu.param1.flags == USE_REG_GEN32
                                 &&  Cpu.param2.flags == USE_DISPLACEMENT32
-                                &&  Cpu.param2.disp32 == GCPtrTpr
+                                &&  Cpu.param2.disp32 == (uint32_t)GCPtrTpr
                                 &&  cbOp == 5)
                             {
                                 /* mov new_reg, uMmioReg */
